ctxutil: Fix context cancellation propagation in go 1.21
Update ctxutil package to correctly handle context cancellation
propagation for go 1.21.

Remove go 1.19 compatability.

Informs #112088.
Needed for #111928.

Release note: None
